The branch's custom-DPI removal already landed on main as the reviewed PR
#892, so the duplicated font/settings/DPI work is dropped in favour of
main's version. The branch's own contributions are reapplied on top of
main's newer code:

- multi-viewport enabled (docking left off) and the Linux/Wayland
  viewport-position and post-drag click-through workarounds
- Linux platform helpers moved into main's RocProfVis::Platform namespace
  alongside the new macOS helpers
- logging initialised before CLI parsing so option diagnostics are captured
- internal build banner drawn on the main viewport
- track resize bars and the sidebar splitter resize via a plain drag
  instead of a payload-less drag-drop source, and track reordering now
  matches on the reorder_request payload type

+++ src/controller/src/rocprofvis_controller_future.cpp
@@ -184,7 +184,7 @@ void Future::ProgressCallback(rocprofvis_db_filename_t db_filename, rocprofvis_d
         {
             future->m_progress_percentage += progress.second;
         }
-        future->m_progress_percentage /= future->m_progress_map.size();
+        future->m_progress_percentage /= static_cast<uint16_t>(future->m_progress_map.size());
         future->m_progress_message = message ? message : "";
     }
 }
